More stuff done
Boss is working fully so that's good, whole game will play through to the end. I need to add a last touch (at least on the subject of bosses) so the boss only spawns when everyone else is gone. If I do that then the boss system is technically done except I want custom models for some bosses I think.