More tweaks
With the help of Eivan and Strabster I've made more changes. Haven't made all the changes mentioned but a decent few. The biggest change is that Strab wants an option to speed up Dialogue, so I'm thinking I rework the system so be like most other games where it ends and you click to continue or whatever