More work on Order Sys
Made some great changes towards the design and the actual functionality of it, I'm getting close actually. Also small change but I made the money icon $ since it was easier to work with. And I fixed the Test Subject to actually use the cost variable so if I re-balance that later it's easy to work with, originally intended to work that way but I set it up incorrectly.