Order Sys Done!!
So the ordering system is complete now, nothing to do but adding some audio to it. I changed the way the level scripts work, now it is a class with a lot of the repetitive actions set up. There are some ways to make it even cleaner but for now this is solid. I also made some smaller tweaks like the journal pages being bigger, reimported some assets, (hopefully) fixed a bug with the charisma hearts, fixed clock, fixed ingredient collisions.